requirements.txt
# Django and Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Environment variables
python-decouple==3.8

# CORS handling
django-cors-headers==4.3.1

# API Documentation
drf-spectacular==0.27.0

# Password validation
argon2-cffi==23.1.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1